Make e2e tests more descriptive
